requirements.txt
selenium==4.15.2
pytest==7.4.3
webdriver-manager==4.0.1
python-dotenv==1.0.0

